move to distribution options component
